Fixed bug in start codon prediction that caused it to prefer the codon before the real start if it was also a possible start codon
Rewrote the criteria for overlapping genes to be > 10% of the db feature length rather than 50 bp.

Modified to stop codon prediction to look for stop codons within 21 bp of the end of the blast hit to deal with cases where the end of the hit is mis-aligned

Modifed BinHits to group together hits within 10 bp of each other.